# [Tentative] Moving slow tokenizer to the Trie world.