--- a/src/transformers/tokenization_utils.py
+++ b/src/transformers/tokenization_utils.py
@@ -52,6 +52,7 @@ TOKENIZER_CONFIG_FILE = "tokenizer_config.json"
 class Trie:
     """
     Trie in Python. Creates a Trie out of a list of words. The trie is used to split on `added_tokens` in one pass
+    Loose reference https://en.wikipedia.org/wiki/Trie
     """
 
     def __init__(self):
@@ -110,21 +111,55 @@ class Trie:
 
         ["[CLS]", " This is a ", "extra_id_100"]
         """
+
+        # indexes are counted left of the chars index.
+        # "hello", index 0, is left of h, index 1 is between h and e.
+        # index 5 is right of the "o".
+
+        # States are going to capture every possible start (indexes as above)
+        # as keys, and have as values, a pointer to the position in the trie
+        # where we're at. This is a partial match for now.
+        # This enables to keep track of multiple matches while we're iterating
+        # the string
+        # If the trie contains, "blowing", and "lower" and we encounter the
+        # string "blower", we need to split into ["b", "lower"].
+        # This is where we need to keep track of multiple possible starts.
         states = {}
+
+        # This will contain every indices where we need
+        # to cut.
         offsets = []
 
+        # This is used by the lookahead which needs to skip over
+        # some text where the full match exceeded the place in the initial
+        # for loop
         skip = None
+        # Main loop, Giving this algorithm O(n) complexity
         for current, c in enumerate(text):
             if skip and current < skip:
                 # Prevents the lookahead for matching twice
                 # like extra_id_100 and id_100
                 continue
+
+            # This will track every state
+            # that stop matching, we need to stop tracking them.
+            # If we look at "lowball", we're going to match "l" (add it to states), "o", "w", then
+            # fail on "b", we need to remove 0 from the valid states.
             to_remove = set()
+            # Whenever we found a match, we need to drop everything
+            # this is a greedy algorithm, it will match on the first found token
             reset = False
+
+            # In this case, we already have partial matches (But unfinished)
             for start, p in states.items():
                 if c in p:
+                    # The current character being looked at has a match within the trie
+                    # update the pointer (it will be stored back into states later).
                     p = p[c]
                     if "" in p:
+                        # This is a final match, we need to reset and
+                        # store the results in `offsets`.
+
                         # Lookahead to match longest first
                         # Important in case of extra_id_1 vs extra_id_100
                         lookahead_index = current + 1
@@ -143,28 +178,49 @@ class Trie:
                             next_c = text[lookahead_index]
                         # End lookahead
 
+                        # Storing and resetting
                         offsets.append(start)
                         offsets.append(end)
                         reset = True
+
+                    # Storing back the new pointer into the states.
+                    # Partial matches got longer by one.
                     states[start] = p
                 else:
+                    # The new character has not match in the trie, we need
+                    # to stop keeping track of this partial match.
+                    # We can't do it directly within the loop because of how
+                    # python iteration works
                     to_remove.add(start)
+
+            # Either clearing the full start (we found a real match)
+            # Or clearing only the partial matches that didn't work.
             if reset:
                 states = {}
             else:
                 for start in to_remove:
                     del states[start]
+
+            # If this character is a starting character within the trie
+            # start keeping track of this partial match.
             if c in self.data:
                 states[current] = self.data[c]
 
+        # We have all the offsets now, we just need to do the actual splitting.
+        # We need to eventually add the first part of the string and the eventual
+        # last part.
         tokens = []
         start = 0
         for end in offsets:
             if start == end:
+                # This might happen if there's a match at index 0
+                # we're also preventing zero-width cuts in case of two
+                # consecutive matches
                 continue
             tokens.append(text[start:end])
             start = end
         if start != len(text):
+            # We're missing the last piece
             tokens.append(text[start:])
         return tokens