--- a/src/transformers/tokenization_utils.py
+++ b/src/transformers/tokenization_utils.py
@@ -65,19 +65,15 @@ class Trie:
 
         This function is idempotent, adding twice the same word will leave the trie unchanged
 
-        >>> trie = Trie()
+        Example::
 
-        >>> trie.add("Hello 友達")
-
-        >>> trie.data
-
-        {"H": {"e": {"l": {"l": {"o": {" ": {"友": {"達": {"": 1}}}}}}}}}
-
-        >>> trie.add("Hello")
-
-        >>> trie.data
-
-        {"H": {"e": {"l": {"l": {"o": {"": 1, " ": {"友": {"達": {"": 1}}}}}}}}}
+            >>> trie = Trie()
+            >>> trie.add("Hello 友達")
+            >>> trie.data
+            {"H": {"e": {"l": {"l": {"o": {" ": {"友": {"達": {"": 1}}}}}}}}}
+            >>> trie.add("Hello")
+            >>> trie.data
+            {"H": {"e": {"l": {"l": {"o": {"": 1, " ": {"友": {"達": {"": 1}}}}}}}}}
         """
         if not word:
             # Prevent empty string
@@ -95,21 +91,15 @@ class Trie:
 
         This trie will match the longest possible word first !
 
-        >>> trie = Trie()
-
-        >>> trie.split("[CLS] This is a extra_id_100")
-
-        ["[CLS] This is a extra_id_100"]
-
-        >>> trie.add("[CLS]")
-
-        >>> trie.add("extra_id_1")
-
-        >>> trie.add("extra_id_100")
-
-        >>> trie.split("[CLS] This is a extra_id_100")
-
-        ["[CLS]", " This is a ", "extra_id_100"]
+        Example::
+            >>> trie = Trie()
+            >>> trie.split("[CLS] This is a extra_id_100")
+            ["[CLS] This is a extra_id_100"]
+            >>> trie.add("[CLS]")
+            >>> trie.add("extra_id_1")
+            >>> trie.add("extra_id_100")
+            >>> trie.split("[CLS] This is a extra_id_100")
+            ["[CLS]", " This is a ", "extra_id_100"]
         """
 
         # indexes are counted left of the chars index.