@Narsil 2021-08-31
@patrickvonplaten It speeds up all the time, but it's only that impressive with adding tons of `added_tokens`. The reason is because of the change in complexity. Even for simple `ByT5` that uses 125 extra `added_tokens` by default, that's already 100x speed (so not really that exotic). 

For "regular" slow tokenizers, with <5 added_tokens the speedup exists but is rather negligible.

Also I've seen 2/3 different issues regarding this. The one reported (https://github.com/huggingface/tokenizers/issues/615) has 8 participants since February, so while not super urgent, there definitely seems to be more than a couple of people doing that.

And always fine adding more documentation, but fwiw it's a pretty standard data structure.