@Narsil 2021-09-03
Hi @LysandreJik Do we have benchmarks  tests anywhere ?

This doesn't fix anything that was broken before, it just make things faster (some usage of the lib was so slow that it was unusable, but it was definitely working).

I could add a test that some tokenization takes too long, but it's always a tricky business to add tests related to timings because it might depend on the hardware running the tests, so it would definitely NOT be a unit test.