@Narsil 2021-09-08
@LysandreJik Ok, I added 1 `common` test which fails only on Canine (`extra_id_1` is not valid over there).

Also added Trie specific tests (the one in the doc basically)