@Narsil 2021-09-08
Thanks @SaulLu , the matching was incorrect in that edge case where some token is rigourously included in another, then we would match the inner token instead of the first match.

The lookahead part got more complex but will now work in that edge case (which is important to at least follow the documentation)

Regarding the idea of `added_tokens` following order, recaping some offline conversation:
- It's doable, but would make code even more complex. We would need to keep track of ranks in the Trie, whenever we have a full match, resolve all partial matches, sort by order and take the highest rank.
- Seems overly complex for what seems to be pathological cases at best, so out of scope of this one.