@Narsil 2021-09-09
Will merge this later today unless there are still some comments (But I feel it's ok in current state)